Fix re-import issue
Previously, Godot would trigger a re-import of the LDtk
levels on every context switch, regardless of whether
the level was modified or not.

This fixes the issue. Root cause: Godot expects that
a resource is saved to the save_path location
passed into the import() function, even if other
files are generated. This is not documented in Godot.